connected show asset components directly to redux store
