+++ react/components/AssetTitle/view.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AssetTitle = ({title}) => {
+  return (
+    <div>
+      <span className="text--large">{title}</span>
+    </div>
+  );
+};
+
+export default AssetTitle;
