+++ react/components/ShowAssetLite/index.js
@@ -1,30 +1,11 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import AssetDisplay from 'containers/AssetDisplay';
+import { connect } from 'react-redux';
+import View from './view';
 
-class ShowLite extends React.Component {
-  render () {
-    const { error, claimData: { name, claimId } } = this.props;
-    return (
-      <div className="row row--tall flex-container--column flex-container--center-center">
-        {error &&
-          <p>{error}</p>
-        }
-        {this.props.claimData &&
-        <div>
-          <AssetDisplay />
-          <Link id="asset-boilerpate" className="link--primary fine-print" to={`/${claimId}/${name}`}>hosted via Spee.ch</Link>
-        </div>
-        }
-      </div>
-    );
-  }
+const mapStateToProps = ({ show }) => {
+  return {
+    name   : show.showAsset.claimData.name,
+    claimId: show.showAsset.claimData.claimId,
+  };
 };
 
-ShowLite.propTypes = {
-  error    : PropTypes.string,
-  claimData: PropTypes.object.isRequired,
-};
-
-export default ShowLite;
+export default connect(mapStateToProps, null)(View);
