+++ react/constants/show_action_types.js
@@ -1,8 +1,10 @@
-export const REQUEST_UPDATE_CHANNEL = 'REQUEST_UPDATE_CHANNEL';
-export const REQUEST_UPDATE_CLAIM = 'REQUEST_UPDATE_CLAIM';
+export const REQUEST_CHANNEL_UPDATE = 'REQUEST_CHANNEL_UPDATE';
+export const REQUEST_CLAIM_UPDATE = 'REQUEST_CLAIM_UPDATE';
+export const REQUEST_ERROR_UPDATE = 'REQUEST_ERROR_UPDATE';
 export const CHANNEL_DATA_UPDATE = 'CHANNEL_DATA_UPDATE';
 export const CHANNEL_CLAIMS_DATA_UPDATE = 'CHANNEL_CLAIMS_DATA_UPDATE';
 export const ASSET_CLAIM_DATA_UPDATE = 'ASSET_CLAIM_DATA_UPDATE';
 export const FILE_REQUESTED = 'FILE_REQUESTED';
-export const FILE_IS_AVAILABLE_UPDATE = 'FILE_IS_AVAILABLE_UPDATE';
+export const FILE_AVAILABILITY_UPDATE = 'FILE_AVAILABILITY_UPDATE';
 export const SHOW_ASSET_ERROR = 'SHOW_ASSET_ERROR';
+export const DISPLAY_ASSET_ERROR = 'DISPLAY_ASSET_ERROR';
