+++ react/sagas/show.js
@@ -1,6 +1,6 @@
 import { call, put, takeLatest } from 'redux-saga/effects';
 import * as actions from 'constants/show_action_types';
-import { updateFileIsAvailable, updateShowAssetError } from 'actions/show';
+import { updateFileAvailability, updateDisplayAssetError } from 'actions/show';
 import { UNAVAILABLE, AVAILABLE } from 'constants/asset_display_states';
 import { checkFileAvailability, triggerClaimGet } from 'api/fileApi';
 
@@ -12,27 +12,27 @@ function* retriveFile (action) {
   try {
     ({ success, message, data: isAvailable } = yield call(checkFileAvailability, name, claimId));
   } catch (error) {
-    return yield put(updateShowAssetError(error.message));
+    return yield put(updateDisplayAssetError(error.message));
   };
   if (success) {
     if (isAvailable) {
-      return yield put(updateFileIsAvailable(AVAILABLE));
+      return yield put(updateFileAvailability(AVAILABLE));
     }
-    yield put(updateFileIsAvailable(UNAVAILABLE));
+    yield put(updateFileAvailability(UNAVAILABLE));
   } else {
-    yield put(updateShowAssetError(message));
+    yield put(updateDisplayAssetError(message));
   }
   // initiate get request for the file
   try {
     ({ success, message } = yield call(triggerClaimGet, name, claimId));
   } catch (error) {
-    return yield put(updateShowAssetError(error.message));
+    return yield put(updateDisplayAssetError(error.message));
   };
   if (success) {
     console.log('/api/glaim-get response:', message);
-    yield put(updateFileIsAvailable(AVAILABLE));
+    yield put(updateFileAvailability(AVAILABLE));
   } else {
-    yield put(updateShowAssetError(message));
+    yield put(updateDisplayAssetError(message));
   }
 }
